BJS2-7260: Analysis of mentoring requests. Fix 1 after review
