+++ src/main/java/faang/school/analytics/dto/MessageEvent.java
@@ -2,19 +2,15 @@ package faang.school.analytics.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
+import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
 
 import java.time.LocalDateTime;
 
-@ToString
 @AllArgsConstructor
 @NoArgsConstructor
-@Getter
-@Setter
 @Builder
+@Data
 public class MessageEvent {
 
     private long actorId;
