+++ src/main/java/faang/school/analytics/listener/EventListener.java
@@ -1,0 +1,39 @@
+package faang.school.analytics.listener;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import faang.school.analytics.dto.MessageEvent;
+import faang.school.analytics.mapper.AnalyticsEventMapper;
+import faang.school.analytics.model.AnalyticsEvent;
+import faang.school.analytics.model.EventType;
+import faang.school.analytics.service.AnalyticsEventService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.redis.connection.Message;
+import org.springframework.data.redis.connection.MessageListener;
+
+import java.io.IOException;
+
+@Slf4j
+@RequiredArgsConstructor
+public abstract class EventListener implements MessageListener {
+
+    private final ObjectMapper objectMapper;
+    private final AnalyticsEventMapper analyticsEventMapper;
+    private final AnalyticsEventService analyticsEventService;
+
+    @Override
+    public void onMessage(Message message, byte[] pattern) {
+        try {
+            String channel = new String(message.getChannel());
+            String body = new String(message.getBody());
+            log.info("Received message from channel {}: {}", channel, body);
+            MessageEvent messageEvent = objectMapper.readValue(body, MessageEvent.class);
+            AnalyticsEvent analyticsEvent = analyticsEventMapper.toEntity(messageEvent);
+            analyticsEvent.setEventType(EventType.MENTORSHIP_REQUESTED);
+            analyticsEventService.saveEvent(analyticsEvent);
+        } catch (IOException e) {
+            log.error(e.getMessage());
+            throw new RuntimeException(e);
+        }
+    }
+}
