+++ src/main/java/faang/school/analytics/listener/MentorshipRequestedEventListener.java
@@ -1,40 +1,16 @@
 package faang.school.analytics.listener;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import faang.school.analytics.dto.MessageEvent;
 import faang.school.analytics.mapper.AnalyticsEventMapper;
-import faang.school.analytics.model.AnalyticsEvent;
-import faang.school.analytics.model.EventType;
 import faang.school.analytics.service.AnalyticsEventService;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.connection.Message;
-import org.springframework.data.redis.connection.MessageListener;
 import org.springframework.stereotype.Component;
-
-import java.io.IOException;
 
 @Slf4j
 @Component
-@RequiredArgsConstructor
-public class MentorshipRequestedEventListener implements MessageListener {
+public class MentorshipRequestedEventListener extends EventListener {
 
-    private final ObjectMapper objectMapper;
-    private final AnalyticsEventMapper analyticsEventMapper;
-    private final AnalyticsEventService analyticsEventService;
-
-    @Override
-    public void onMessage(Message message, byte[] pattern) {
-        try {
-            String channel = new String(message.getChannel());
-            String body = new String(message.getBody());
-            log.info("Received message from channel {}: {}", channel, body);
-            MessageEvent messageEvent = objectMapper.readValue(body, MessageEvent.class);
-            AnalyticsEvent analyticsEvent = analyticsEventMapper.toEntity(messageEvent);
-            analyticsEvent.setEventType(EventType.MENTORSHIP_REQUESTED);
-            analyticsEventService.saveEvent(analyticsEvent);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
+    public MentorshipRequestedEventListener(ObjectMapper objectMapper, AnalyticsEventMapper analyticsEventMapper, AnalyticsEventService analyticsEventService) {
+        super(objectMapper, analyticsEventMapper, analyticsEventService);
     }
 }
