+++ src/main/java/faang/school/analytics/service/AnalyticsEventService.java
@@ -3,9 +3,11 @@ package faang.school.analytics.service;
 import faang.school.analytics.model.AnalyticsEvent;
 import faang.school.analytics.repository.AnalyticsEventRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class AnalyticsEventService {
@@ -14,6 +16,8 @@ public class AnalyticsEventService {
 
     @Transactional
     public void saveEvent(AnalyticsEvent analyticsEvent) {
+        log.info("Save event with actor ID: {}, receiver ID: {}, event type: {}",
+                analyticsEvent.getActorId(), analyticsEvent.getReceiverId(), analyticsEvent.getEventType().name());
         analyticsEventRepository.save(analyticsEvent);
     }
 }
